Move non-helpful stuff into their own file
